triketime-backend/requirements.txt
Flask
gunicorn
python-dotenv
google-cloud-firestore
flask-cors
